rename input_product_readers/_utils.py
