+++ src/nisarqa/input_product_readers/l1_radar_base.py
@@ -8,7 +8,7 @@ import shapely
 
 import nisarqa
 
-from ._product_reader_utils import (
+from ._utils import (
     _get_dataset_handle,
     _get_fill_value,
     _get_or_create_cached_memmap,
